requirements.txt
# requirements.txt

Flask
pandas
google-generativeai
Flask-Cors 